requirements.txt
# Core Data Science
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Statistics & Modeling
statsmodels==0.14.1
scikit-learn==1.3.2

# Dashboard
streamlit==1.29.0
plotly-express==0.4.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
pytest==7.4.3

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.27.1